lsp: set sortText for Element properties
If the sortText isn't there, then label is used to do the sorting [1]
which results in a poor auto-complete experience when trying to see what
properties are available for a component or just browsing the api of a
component.

This patch sets sortText on properties such that they will always be on
top.

[1]: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#completionItem