Fix initial rendering of the live-preview
The hack to sets the preview area's width and height through the init callback was run _after_ rendering the component container and during evaluation
of the dirty tracker. The latter sets itself to dirty = false after evaluation, so it's never dirty despite
the width and height having changed.

As a workaround for 1.4.1, move the init callback hack to before render() is called.