Fix integrity of FilterModel and ReverseModel when removing or adding rows
The ReverseModel had a bug that it didn't compute the index correctly
when removing several rows at the time

The FilterModel didn't adust its internal mapping if the row that was
removed was filtered out.

That caused a bug that can be reproduced in the todo demo:
Check "hide done items" and "Sort by name", then click "Remove Done
Items", then add a new entry and it panicked.