lsp: Simplify sending messages from preview to lsp
... in native mode,

Have the channel between the two transport
`PreviewToLspMessages` directly instead of wrapping
those into `Notifications` of the language server
protocol.

Unformtunately the notification code eneds to stay in
place, as that is what the WASM preview uses to talk
to a native LSP.

I did nto unify wasm and native handling of the
`PreviewToLspMessage`s: The different async runtimes
we need to integrate with got into the way.