C++ package: use Qt on Linux, not on Windows/Mac
Commit 4b4a7d6ff2d1927ca53bd5be7fdf4f5bf15a8e63 attempted to remove the
Qt from the C++ package on Windows, but infortunately, it removed the
setup of Qt on Linux instead of Windows.

As a result, 1.4.x binary packages on Linux are not build with Qt, and
on Windows are build with Qt but no longer ship Qt so are broken

In the master branch we no longer use Qt by default already on Windows
even if it is installed.
But this patch restore the fact that the binary package are built with
Qt on Linux