live preview: Use selection based on nodes
Now that we keep information on all nodes, use that to do more
"fine grained" selection: We select based on nodes, not elements in
the element tree now.

This assumes that a source locartion based on file path and offset in
that file is enough to define the nodes of an element.