Fix cross build of rust-skia in cross docker containers
Since skia-bindings-0.71.0, bindings.cpp includes SkFontMgr_fontconfig.h, which in turn includes fontconfig/fontconfig.h.

In cross docker containers, the fontconfig headers are installed in /usr/include - there's no sysroot.

For the actual build of bindings.cpp, this is fine - gcc will search in /usr/include and find it.

For the bindgen phase, certain clang versions do not look in -I/usr/include by default when a --target is present - generally a sensible approach probably. However in this very setup, it's okay, so teach bindgen about it by setting BINDGEN_EXTRA_CLANG_ARGS_target. Typically that's used to pass for example `--sysroot=`, but we're using it to explicitly add /usr/include.