fmt: Fix formatting of long chain of if-else
We would end up adding too many new line.

The thing is that the `else` is inside the Expression node and cannot be
detected in format_conditional_expression, so the code was dead.
Instead, rely on the inner if to be formatted correctly, but clean the
whitespace_to_add to collapse new lines