live preview: Fix WASM preview talking to native LSP
We created a `ServerNotifier`, cloned that and passed one copy
on to the `CompilerConfiguration` so that imports would notify the
preview.

Unfortunately that does not work: That clone never gets told to
switch over to the WASM preview and will always try to update the native
preview instead.

So make sure the information which preview is used gets
synchrnoized to all clones using an atomic bool and an Arc...

Fixes: #4764