Improve rendering performance of the energy monitor
The use of gradients in the design to implement bar charts prevents Skia from batching, which means
that each bar is a separate draw call.

Since we barely change the model data, use a cache-rendering-hint to batch the bar charts by hand into a texture.

Similarly, when scrolling on the main screen, we don't need to re-draw each
page individually until we're selecting.

This boosts the scrolling performance from 13fps to 30fps on an stm32mp157.

See #4944 for separate docs about this.