Fix WASM Loading Issue Due to Library Name Change (#4990)
Following the recent changes in #4765, the library name was updated from printerdemo to printerdemo_lib. This modification has inadvertently affected the naming of the JavaScript file generated by wasm-pack, leading to printerdemo.js being renamed to printerdemo_lib.js. Consequently, the existing reference in the index.html file became outdated, causing the WASM module to fail to load correctly.

This PR addresses the issue by updating the reference in index.html, ensuring that the page can properly load and execute the WASM module.