Skia: Fix quality of control rendering when using `cache-rendering-hint: true;`
By default, render layers with smooth scaling, just like we do also in `Image` by default.

The easiest way to reproduce the issue is to use `cache-rendering-hint: true;` on an `Image` and then scale it: It'll appear as if `image-rendering: pixelated;` is used, which it shouldn't.

`cache-rendering-hint: true;` on an `Image` doesn't make much sense by itself, but it should be a no-op.