Fix a warning in default Python build about unused create_window() function in EventLoopInterface
Yes, the fix for that is here, instead of cluttering #cfg's into the winit backend. This seemingly unused function is not needed by femtovg on desktop,
merely by skia (not enabled by default) and the software renderer.

In the Slint Rust crate, the software renderer is enabled by default - which is why this doesn't trigger a warning. But for the interpreter it was missing.