Zephyr: individually render the dirty sub-regions
The software renderer now provides a list of regions which have
changed.  Convert the regions to big-endian and render them
individually to prevent the non-dirty regions from being converted
away from big-endian.  Unfortunately, this still means we convert any
intersection between region more than once but this will be tackled as
a follow-up commit, as in principle this affects more than just the
Zephyr demo.