Zephyr: active animations only cause a sleep on the native simulator
The Zephyr POSIX architecture used by the native simulator is unable
to interrupt a busy thread [1]. Therefore we must sleep even when
there are active animations to allow other threads to progress,
otherwise we end up in an infinite loop. This limitation does not
apply to real hardware, where we can simply continue the event loop.

[1] https://docs.zephyrproject.org/3.7.0/boards/native/doc/arch_soc.html#important-limitations