Rename the weather-demo library name
The reason why we need to have a `_lib` prefix in the printerdemo and
the too example is because the binary name must be different from the
library name. But in the case of the weather-demo, since there is a
dash, it doesn't need the _lib suffix to be different.

Having a different name than the package name doesn't work with the
`xbuild` tool: fixes #5731
(we use cargo apk for our own build which doesn't have this limitation
anyway)