Zephyr: CI uses appropriate toolchain for the target
The simulator build requires the nightly rust toolchain, and
occasionally fails due to upstream changes to the nightly toolchain.

The hardware build does not require the nightly rust toolchain, but
still fails due to upstream nightly toolchain changes.

Move the hardware build to the stable toolchain so that it is not
broken unnecessarily.

Additionally, turn off `fail-fast` so that the nightly failures don't
affect the stable builds.

Fixes #5812.