compiler: file lookup returns contents and version
Have the `CompilerConfiguration::open_import_fallback` return
both the contents and the version of a file.

The default implementation set via the setters in public API
still take a callback that returns the fiole contents only and
the version will default to `None`.

This is to keep public API unchanged while letting people
opting in to internal interfaces set file versions.