core: Do not panic when popups try to find their parent
Popups may have normal `Item`s as parrent, while all the other
components always have a `DynamicTree` as a parent. So do not
panic when some component ends up with an `Item` as a parent.

Just return `None` instead of panicing.

We probably need to make the focus-handling more aware of
popups eventually, but this stops the panics.

Fixes: #5826, #5830