compiler: Treat ASCII SOT and EOT as whitespace
This is so we can use the ASCII Start-of-Text and End-of-Text
control characters as markers in files containing blocks of Slint
code embedded into other things.

SOT and EOT are ASCII control characters, so they are one byte and
thus will basically fit into whatever markup is used to separate
Slint code form whatever else surrounds it.