live-preview: Use TextRange and TextSize in property information
The Properties are no longer sent out straight to the editor, so
we do not need to convert our internal `TextRange` and `TextSize`
to `lsp_types::Range` and `lsp_types::Position`.

We are noweadys turning that straight back into offsets -- which is
just another name for `TextSize` (or pairs of offsets).