vscode extension: make sure the "Slint LSP" output always exist
It is used to show the output of `debug()`.
User don't find this output if it doesn't exist before there is any
debug output. So make sure it always exist.

(Note: this is not for the browser version, since then we use
console.log instead of the "Slint LSP" output from stderr)