live-preview: Fix wrong highlight color for layouts
Elements are specified by theit path and start offset. So
request a highlight for that position, not for some random
position somewhere inside the Element to highlight.

That way looking up whether the element is actually
a layout or not does work -- and there is no fallback to
the "it's not a layout" case.