live-preview: Fix size being displayed wrongly
Stop abusing the init callback, abuse the changed callback instead ;-)

Also merge the initial size calculation into one place. This fixes the
preview showing the wrong sizes. It does introduce a bit of flicker though
as the UI is rendered in the wrong state once and then immediently corrects
itself.

That is annoying, but strictly better than staying in the wrong state till
the user does something.

Closes: #5954