Slintpad: add an about dialog
Back in Slint 1.6, slintpad had a "welcome" page that explained what it is.
But the welcome page was removed in commit 588ee3e4740eda68b36ff2f55323b2f2522a42b2
Now it's hard for someone who click on a link to slintpad.com to know
what it is and get a link to Slint.

Add an "about" entry in the menu that shows the previous welcome page as
a dialog