live-preview: Sort properties and split them up into groups
I was stupid this morning: The properties need to be sorted.

So I need to group the properties into groups and I also want
to be able to define in which order the groups will be listed
in the UI. So add a `priority` in addition to the group and use
group name and priority to sort properties into groups.

The test changes since the order changed and it tries to insert
based on previous/later properties.