live-preview: Filter out properties that make no sense
Filter out properties related to layouts that make no sense
for the current element. These are x/y coordinates and the
grid layout related properties.

It's tricky to do: The properties are looked up in the
context of the unoptimized DocumentCache, so that
(most) properties are visible. The layout information is
not available there though as base types are not resolved yet.

So the layout information is taken from the preview data.