vscode: Request interesting buffers only
Ask VSCode to only send files stored on disc to the LSP,
not random other kinds of buffers. The LS can not make
sense of anything else anyway as it can nto get to
the resources stored next to buffer it works on.

For WASm we can load more kinds of buffers. Whitelist a range
of possibnle buffer schemes, incl. `vscode-vfs` which is used
on `github.dev`.

This does stop VSCode from asking the LS to process random
other buffers like those containing a git diff of a slint file.

Fixes: #6375