Fix a case where a timer would end up in the timeout sorted list twice
That in turn can cause panics.

Suppose there are two timers: One expiring sooner and one later. During timer activation of the sooner timer, the second timer is restarted. That inserts the timer into the active timers list (correct), but when timers_to_process reaches this timer, it unconditionally tried to register it as active again.