When timers not stopping when stop() is called from activation callback of an earlier timer
During timer activation, the active_timers list did not contain the future timer yet, so calling stop() would not succeed in removing from active_timers.

Fix this by making sure active_timers is up-to-date in terms of timers to be activated in the future by the time we start invoking the activation callbacks.