C++: Avoid undefined behaviour when casting float to int
When the truncated float can't be represented as int, it's undefined behaviour to cast. Work around this by clamping.

Unfortunately this has the side-effect of eliminating an implicit sign conversion that causes signed vs. unsigned comparison warning, when comparing images. The compiler is told that the width/height is Type::Int32, when we return it as uint32_t. This is eliminated with a private helper function.